nl2or_agent/pyproject.toml
[project]
name = "nl2or-agent"
version = "0.1.0"
description = "NL2OR Agent — 将自然语言运筹学问题转化为数学模型并自动求解"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "minds-hamlet>=0.1.3",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "gurobipy",
    "litellm>=1.83.11",
    "scipy>=1.15.3",
    "pulp>=3.3.0",
    "numpy>=2.2.6",
    "pandas>=2.3.3",
]

[project.optional-dependencies]
# Web UI（基于 hamlet 内置的 GradioUI）
web = []

[project.scripts]
nl2or = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.4.0",
]

[tool.ruff]
line-length = 100
target-version = "py310"
